commits to notes
